requirements.txt
flask
pandas
scikit-learn
matplotlib
requests
python-dotenv